ref: consolidate duplicated cache-write, repl-io, and compiler helper logic
Extract five verbatim clones found by duplicate detection:

- LockedPhpCacheWriter: shared flock'd var_export cache-file write for
  PhpNamespaceCache and PhpScanIndexCache (merge-under-lock preserved)
- ReplOutputTrait: shared error/stdout writing half of the REPL IO
  implementations (system + fallback)
- LoweredMatchEmitterTrait: shared PHP match emission for the cond/case
  lowering in IfEmitter and LetEmitter
- AnalyzeBindingsTrait: shared let/loop binding-vector analysis
  (gensym shadow + sequential env threading)
- NsStateDefinitionsEmitterTrait: shared *file*/*ns* addDefinition
  emission for ns and in-ns forms

Emitted PHP and cache-file semantics are byte-identical.
