+++ src/php/Compiler/Domain/Emitter/OutputEmitter/NodeEmitter/IfEmitter.php
@@ -15,6 +15,7 @@ use function assert;
 
 final class IfEmitter implements NodeEmitterInterface
 {
+    use LoweredMatchEmitterTrait;
     use WithOutputEmitterTrait;
 
     public function emit(AbstractNode $node): void
@@ -195,23 +196,7 @@ final class IfEmitter implements NodeEmitterInterface
             return false;
         }
 
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitContextPrefix($env, $loc);
-        $this->outputEmitter->emitStr('match (', $loc);
-        $this->outputEmitter->emitNode($shape['init']);
-        $this->outputEmitter->emitStr(') { ', $loc);
-
-        foreach ($shape['arms'] as $arm) {
-            $this->outputEmitter->emitLiteral($arm['key']);
-            $this->outputEmitter->emitStr(' => ', $loc);
-            $this->outputEmitter->emitLiteral($arm['expr']);
-            $this->outputEmitter->emitStr(', ', $loc);
-        }
-
-        $this->outputEmitter->emitStr('default => ', $loc);
-        $this->outputEmitter->emitLiteral($shape['fallback']);
-        $this->outputEmitter->emitStr(' }', $loc);
-        $this->outputEmitter->emitContextSuffix($env, $loc);
+        $this->emitLoweredMatch($shape, $env, $node->getStartSourceLocation());
 
         return true;
     }
