+++ src/php/Run/Domain/Repl/ReplCommandSystemIo.php
@@ -4,17 +4,16 @@ declare(strict_types=1);
 
 namespace Phel\Run\Domain\Repl;
 
-use Phel\Shared\Exceptions\AbstractLocatedException;
 use Phel\Shared\Facade\ApiFacadeInterface;
 use Phel\Shared\Facade\CommandFacadeInterface;
-use Phel\Shared\Parser\ReadModel\CodeSnippet;
 use Phel\Shared\ScalarCoercion;
-use Throwable;
 
 use function count;
 
 final readonly class ReplCommandSystemIo implements ReplCommandIoInterface
 {
+    use ReplOutputTrait;
+
     public function __construct(
         private string $historyFile,
         private CommandFacadeInterface $commandFacade,
@@ -47,49 +46,6 @@ final readonly class ReplCommandSystemIo implements ReplCommandIoInterface
         }
 
         return $line;
-    }
-
-    public function writeStackTrace(Throwable $e): void
-    {
-        $this->writeln($this->commandFacade->getStackTraceString($e));
-    }
-
-    public function writeReplError(Throwable $e): void
-    {
-        $this->writeln($this->errorFormatter->render($e));
-    }
-
-    public function writeLocatedException(AbstractLocatedException $e, CodeSnippet $codeSnippet): void
-    {
-        $this->writeln($this->commandFacade->getExceptionString($e, $codeSnippet));
-    }
-
-    /**
-     * @psalm-taint-escape html
-     * @psalm-taint-escape has_quotes
-     *
-     * @psalm-suppress TaintedHtml
-     * @psalm-suppress TaintedTextWithQuotes
-     */
-    public function write(string $string = ''): void
-    {
-        /** @psalm-suppress TaintedHtml */
-        /** @psalm-suppress TaintedTextWithQuotes */
-        echo $string; // phpcs:ignore
-    }
-
-    /**
-     * @psalm-taint-escape html
-     * @psalm-taint-escape has_quotes
-     *
-     * @psalm-suppress TaintedHtml
-     * @psalm-suppress TaintedTextWithQuotes
-     */
-    public function writeln(string $string = ''): void
-    {
-        /** @psalm-suppress TaintedHtml */
-        /** @psalm-suppress TaintedTextWithQuotes */
-        echo $string . PHP_EOL; // phpcs:ignore
     }
 
     public function isBracketedPasteSupported(): bool
